requirements.txt
google-api-python-client
futsu
google-auth-httplib2
google-auth-oauthlib
oauth2client
